added basic multiconsole support

+++ app/x32Types/mutegroup.cpp
@@ -1,6 +1,6 @@
 #include "mutegroup.h"
 
-Mutegroup::Mutegroup(X32ConsoleAbstract *console, qint16 number, bool state, QObject *parent) : QObject(parent),
+Mutegroup::Mutegroup(X32ConsoleAbstract *console, quint8 number, bool state, QObject *parent) : QObject(parent),
     console(console), refreshTimer(new QTimer()), number(number), state(state)
 {
     connect(refreshTimer, SIGNAL(timeout()), this, SLOT(timedRefresh()));
